Videos UI: Accept a full footer component as a prop.
