+++ client/components/videos-ui/modal-footer-bar.scss
@@ -1,0 +1,18 @@
+@import '@wordpress/base-styles/breakpoints';
+@import '@wordpress/base-styles/mixins';
+
+.videos-ui__modal-footer-bar {
+	background-color: #101517;
+	border-top: 1px solid rgba( 255, 255, 255, 0.1 );
+	margin-top: auto;
+	padding-top: 16px;
+	position: fixed;
+	bottom: 0;
+	width: 100%;
+}
+
+@include break-small {
+	.videos-ui__modal-footer-bar {
+		display: none;
+	}
+}
