+++ client/my-sites/customer-home/cards/education/blogging-quick-start/blogging-quick-start-modal.jsx
@@ -1,5 +1,6 @@
 import { BlankCanvas } from 'calypso/components/blank-canvas';
 import VideosUi from 'calypso/components/videos-ui';
+import ModalFooterBar from 'calypso/components/videos-ui/modal-footer-bar';
 import ModalHeaderBar from 'calypso/components/videos-ui/modal-header-bar';
 
 import './style.scss';
@@ -11,7 +12,10 @@ const BloggingQuickStartModal = ( props ) => {
 		isVisible && (
 			<BlankCanvas className={ 'blogging-quick-start-modal' }>
 				<BlankCanvas.Content>
-					<VideosUi headerBar={ <ModalHeaderBar onClose={ onClose } /> } />
+					<VideosUi
+						headerBar={ <ModalHeaderBar onClose={ onClose } /> }
+						footerBar={ <ModalFooterBar onBackClick={ onClose } /> }
+					/>
 				</BlankCanvas.Content>
 			</BlankCanvas>
 		)
